requirements.txt
streamlit
PyPDF2
cassio
langchain
astrapy
cassandra-driver
openai
python-dotenv